\name{dataSets}


\alias{dataSets}

\alias{GCCRET}


\title{Assets Data Sets}


\description{
  
    Example data sets for portfolio optimization.
    
}
    

\usage{
GCCRET
}
  


%\details{
%}


\value{
  
    an object of class \code{"timeSeries"}.
  
}


\references{


Wuertz, D., Chalabi, Y., Chen W., Ellis A. (2009);
    \emph{Portfolio Optimization with R/Rmetrics}, 
    Rmetrics eBook, Rmetrics Association and Finance Online, Zurich.
    
}


\examples{
## Gulf Country Indices -
   head(GCCRET)
}


\keyword{data}

